requirements.txt
Jinja2~=2.11.1
Markdown~=3.2.1
markdown-full-yaml-metadata~=2.0.1
python-dateutil~=2.8.1
